Significantly decrease file size of hero background image
This fix reduces file size of hero-section-bg2.jpg from over 5MB to 401KB at the 4k resolution.
By replacing the line in style.css which sets the background-image property with a script at
the bottom of index.html, users do not have to download an image way too big for their display
(eg 768px smartphone), but rather, an image with a size close to their display resolution. For
smartphone users (especially those on 3G or otherwise slow network speeds), this can result in
a file size decrease of 99%. I've tested on Chromium (Edge), but if the code works fine in all
browsers I can turn it into a function and call it on the other large images in index.html